Refactor : 디렉토리를 이용해 컴포넌트 분리
컴포넌트의 개수가 너무 많아져서 관련 있는 것들끼리 묶어놨습니다.
