+++ src/components/Party/PartyDetailHeader.jsx
@@ -2,10 +2,10 @@ import styled from '@emotion/styled';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
-import { COLORS, HOVER_CURSOR_PONTER, SIZES } from '../styles/constants';
-import { convertStatus, convertPlace, convertDate } from '../utils/converter';
-import { icons, images } from '../assets/assets';
-import plainButton from '../styles/plainButton';
+import { COLORS, HOVER_CURSOR_PONTER, SIZES } from '../../styles/constants';
+import { convertStatus, convertPlace, convertDate } from '../../utils/converter';
+import { icons, images } from '../../assets/assets';
+import plainButton from '../../styles/plainButton';
 
 const Container = styled.div`
   display: flex;
