+++ src/pages/IndexPage.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import Header from '../components/Header';
 import Index from '../components/Index';
 import Main from '../components/Main';
-import Parties from '../components/Parties';
+import Parties from '../components/Party/Parties';
 import Footer from '../components/Footer';
-import PartyHeader from '../components/PartyHeader';
+import PartyHeader from '../components/Party/PartyHeader';
 import useParty from '../hooks/useParty';
 
 function IndexPage() {
